go! aku!
A rewrite of the shapeshifting master of evil who likes to shout things
for your auditory gratification. This time it's in a language with no
generics.

So far he knows how to connect to discord and track people's voice
states so that in the future he'll be able to yell things when people
transition from being nowhere to being somewhere. He also takes more
careful notes about his yelling and such than his predecessor for the
benefit of posterity.